requirements.txt
flask
flask-sqlalchemy
flask-migrate
werkzeug
flask-login
google-cloud-speech
